requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
sentence-transformers
scikit-learn
numpy
pandas
requests
PyPDF2
bcrypt